Replace Repository ABC boilerplate by a Protocol
